requirements.txt
django
gunicorn
django-heroku
grpcio-tools
git+https://github.com/deepset-ai/haystack.git
farm-haystack
boto3
requests
